+++ R/map_env_dist.R
@@ -53,7 +53,7 @@
 #' sp_pa_2
 #'
 #' # Measure environmental distance between presences and projection data
-#' clrs = c("#000033", "#1400FF", "#C729D6", "#FF9C63", "#FFFF60")
+#' clrs <- c("#000033", "#1400FF", "#C729D6", "#FF9C63", "#FFFF60")
 #'
 #'
 #' # Domain